+++ CHANGELOG.md
@@ -56,6 +56,7 @@ CHANGELOG
 - `intelmq.bots.parsers.anubisnetworks`: Update parser with new data format.
 - `intelmq.bots.parsers.bambenek`: Add new feed URLs with Host `faf.bambenekconsulting.com` (#1525, PR#1526).
 - `intelmq.bots.parsers.abusech.parser_ransomware`: Removed, as the feed is discontinued (#1537).
+- `intelmq.bots.parsers.nothink.parser`: Removed, as the feed is discontinued (#1537).
 
 #### Experts
 - `intelmq.bots.experts.cymru_whois.lib`: Fix parsing of AS names with unicode characters.
@@ -75,6 +76,7 @@ CHANGELOG
   - Remove unavailable Fraunhofer DDos Attack feeds.
   - Remove unavailable feed Abuse.CH Ransomware Tracker (#1537).
   - Update information on Bambenek Feeds, many require a license now (#1525).
+  - Remove discontinued Nothink Honeypot Feeds (#1537).
 - Developers Guide: Fix the instructions for `/opt/intelmq` file permissions.
 
 ### Packaging