+++ NEWS.md
@@ -38,6 +38,10 @@ The Fraunhofer DDoS Attack feed previously available under `https://feed.caad.fk
 #### Bambenek Feeds
 Many Bambenek feeds require a license now and URLs have changed. See https://osint.bambenekconsulting.com/feeds/ for more information. The `intelmqctl upgrade-config` command also warns if you have previously documents feeds in use and migrates the URL for the DGA domain feed.
 
+#### Nothink Feeds and Parser
+All Nothink Honeypot feeds have been discontinued and current the data available covers the time until 2019. The Nothink Parser has been removed.
+The `intelmqctl upgrade-config` command warns if you have these feed and the bot in use.
+
 ### Libraries
 
 ### Postgres databases